Created an explicit local copy of the mdir's mid in lfsr_mdir_commit
Knowing C's issues with pointer aliasing, I was wondering if this might
save some code cost by avoiding unnecessary indirect loads of the mid.

But, as is often the case, the compiler is smarter than it first appears:

            code          stack
  before:  21052           1744
  after:   21048 (-0.0%)   1744 (+0.0%)

Still, sometimes an optimization is better when written out explicitly,
so I'll keep this for now.