Moved around some bits in internal tags to make space for a deferred bit
This is in order to support deferred-inlined files, which involves
intertwining secondary trees into an rbyd. In order to know which trunks
go to which trees, we need an additional bit to indicate if a tag is on
the primary tree or a secondary tree.

We were using pretty much all of our tag bits, but the rm and valid
bits can be combined. They more-or-less serve the same purpose.

New tag modes:

  v000tttt 0ttttttt - normal tags
  v001tttt 0ttttttt - deferred tags
  v010tttt 0ttttttt - checksum tags
  v1dckkkk 0kkkkkkk - alt tags
  ^'+''-----+-----'
  '-|-------|- valid bit
    '-------|- tag mode
            '- tag type/key

Note that once we have a trunk, we don't need this deferred bit to
traverse the rbyd. This is why we can get away with using a mode bit
that would normally collide with the alt tag's encoding.

Also tweaked lfsr_rbyd_appendattr so GROW tags don't need to set the rm
bit anymore. This is just an internal usability thing.