Switched to using mid to tell which files belong in a compaction
This avoids the previous issues with block state for null inlined data,
and we're already testing the rid anyways for splits.

In theory we don't need the block for inlined data at all, but it is
convenient as it allows us to use the existing internal rbyd/data APIs
without needing to move data around. Though it may be worth looking into
alternative layouts at some point.