Deduplicated shrub updates into lfsr_file_carveinlined
lfsr_file_carveinlined writes data into a shrub, while handling both the
carving logic of data we might be overlapping, and any hole logic we
need to fill out the tree.

This provides a nice, relatively simple but flexible, operation for all
shrub updates:

  static int lfsr_file_carveinlined(lfs_t *lfs, lfsr_file_t *file,
          lfs_off_t pos, lfs_off_t weight, lfs_soff_t delta,
          lfsr_data_t data);

I'm quite happy with how these internal carveinlined/carvebtree
functions are coming together. It's nice to have all of that logic in
one place, even if it's a bit complex.